pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mvola-api-lib"
version = "2.0.0"
description = "Une bibliothèque Python robuste et sécurisée pour l'intégration de l'API de paiement MVola"
readme = "README.md"
authors = [
    {name = "Niainarisoa", email = "niainarisoa.mail@gmail.com"}
]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]
keywords = ["mvola", "payment", "api", "madagascar", "fintech", "mobile money", "security"]
dependencies = [
    "requests>=2.25.0,<3.0.0",
    "python-dotenv>=0.19.0,<2.0.0",
    "urllib3>=1.26.0,<3.0.0",
]
requires-python = ">=3.9"

[project.urls]
"Homepage" = "https://github.com/Niainarisoa01/Mvola_API_Lib"
"Bug Tracker" = "https://github.com/Niainarisoa01/Mvola_API_Lib/issues"
"Documentation" = "https://niainarisoa01.github.io/Mvola_API_Lib"
"Changelog" = "https://github.com/Niainarisoa01/Mvola_API_Lib/blob/main/docs/changelog.md"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12.0",
    "black>=21.5b2",
    "isort>=5.9.1",
    "flake8>=3.9.2",
    "bandit>=1.7.0",
    "safety>=2.0.0",
]
docs = [
    "mkdocs>=1.3.0",
    "mkdocs-material>=8.2.8",
    "mkdocstrings>=0.18.0",
    "mkdocstrings-python>=0.7.1",
    "mike>=1.1.2",
]
examples = [
    "flask>=2.0.0",
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests", "scripts"]
skips = ["B101"]

[tool.hatch.build.targets.wheel]
packages = ["mvola_api"]